feat (logging): add an stdout Backend with json formatting
